// contracts/amm/rewards/schema/init.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HumanAddr = string;

export interface Init {
  admin?: HumanAddr | null;
  config: RewardsConfig;
  [k: string]: unknown;
}
/**
 * Reward pool configuration
 */
export interface RewardsConfig {
  bonding?: number | null;
  lp_token?: ContractLink | null;
  reward_token?: ContractLink | null;
  reward_vk?: string | null;
  timekeeper?: HumanAddr | null;
  [k: string]: unknown;
}
/**
 * Info needed to talk to a contract instance.
 */
export interface ContractLink {
  address: HumanAddr;
  code_hash: string;
}